+++ app/mqresources/mqutils.py
