+++ app/translation_service/translation_service.py
@@ -1,0 +1,14 @@
+import os, os.path, logging
+from mqresources import mqutils
+
+logfile=os.getenv('LOGFILE_PATH', 'hdc3a_transfer_service')
+loglevel=os.getenv('LOGLEVEL', 'WARNING')
+logging.basicConfig(filename=logfile, level=loglevel)
+
+def translation_data_structure():
+    return True
+
+
+    
+        
+
