+++ docker-compose-local.yml
@@ -4,8 +4,8 @@ version: '3.8'
 
 services:
 
-  mqutils:
-    container_name: 'mqutils'
+  translation_service:
+    container_name: 'translation_service'
     build:
       context: '.'
       dockerfile: 'Dockerfile'
